# Remove excessive graph nodes iteration during ConnectedGraph construction