--- a/TrainingExtensions/torch/test/python/test_connectedgraph.py
+++ b/TrainingExtensions/torch/test/python/test_connectedgraph.py
@@ -791,37 +791,37 @@ class TestConnectedGraphUtils(unittest.TestCase):
         # 1) aimet_modules.Add()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.Add(), dummy_input)
-        nodes =  ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes =  list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 2) aimet_modules.Subtract()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.Subtract(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 3) aimet_modules.Multiply()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.Multiply(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 4) aimet_modules.Divide()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.Divide(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 5) aimet_modules.MatMul()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.MatMul(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 6) aimet_modules.Concat()
         dummy_input = (torch.randn(1, 3, 4, 4), torch.randn(1, 3, 4, 4))
         trace = torch.jit.trace(aimet_modules.Concat(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
     @pytest.mark.cuda
@@ -836,7 +836,7 @@ class TestConnectedGraphUtils(unittest.TestCase):
 
         dummy_input = torch.randn(1, 3, 4, 4)
         trace = torch.jit.trace(CustomModule().cuda(), dummy_input.cuda())
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         # mulitply, softplus and relu ops.
         # detach is considered as passthrough op.
         assert len(nodes) == 3
@@ -854,7 +854,7 @@ class TestConnectedGraphUtils(unittest.TestCase):
         dummy_bias = torch.randn((32,))
         dummy_input = (dummy_input, dummy_weight, dummy_bias)
         trace = torch.jit.trace(CustomModule(), dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
     def test_find_nodes_in_forward_pass_for_torch_nn_module(self):
@@ -865,21 +865,21 @@ class TestConnectedGraphUtils(unittest.TestCase):
         conv = torch.nn.Conv2d(3, 3, 2).eval()
         print(conv.__class__)
         trace = torch.jit.trace(conv, dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 2) ReLU
         dummy_input = torch.randn(1, 3, 4, 4)
         relu = torch.nn.ReLU(inplace=True).eval()
         trace = torch.jit.trace(relu, dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
         # 3) BatchNorm2d
         dummy_input = torch.randn(1, 3, 4, 4)
         bn = torch.nn.BatchNorm2d(3).eval()
         trace = torch.jit.trace(bn, dummy_input)
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(trace))
         assert len(nodes) == 1
 
     def test_find_nodes_in_forward_pass_for_unused_module(self):
@@ -909,12 +909,12 @@ class TestConnectedGraphUtils(unittest.TestCase):
 
         # Conv2 is unused in forward pass.
         conv2_trace = getattr(trace, "conv2")
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(conv2_trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(conv2_trace))
         assert len(nodes) == 0
 
         # Conv1 and Conv3 are used in forward pass.
         conv1_trace = getattr(trace, "conv1")
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(conv1_trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(conv1_trace))
         assert len(nodes) == 1
 
     def test_find_nodes_in_forward_pass_for_undefined_graph(self):
@@ -927,7 +927,7 @@ class TestConnectedGraphUtils(unittest.TestCase):
 
         inner_seq_trace = getattr(trace, "inner_seq")
         bn_trace = getattr(inner_seq_trace, "1")
-        nodes = ConnectedGraph._find_aten_nodes_in_forward_pass(bn_trace)
+        nodes = list(ConnectedGraph._find_aten_nodes_in_forward_pass(bn_trace))
         assert len(nodes) == 0
 
         with pytest.raises(RuntimeError):