## Main Changes
Optimized `ConnectedGraph.__init__`

### Benchmarks
* model: Unet + LoRA (1.3B parameters) (thanks @quic-mangal)
* GPU: RTX 2080

|                                                                            |     current     |   this PR    |
|----------------------------------------------------|------------|------------|
| `QuantizationSimModel.__init__`                        |    ~138s    |    **~28s**  |
| └ `ConnectedGraph.__init__`                              |    ~133s    |    **~23s**   |
| &nbsp;&nbsp;&nbsp;&nbsp;└ `torch.jit.trace`  |    ~20s      |    ~20s     |
| &nbsp;&nbsp;&nbsp;&nbsp;└ `<others>`        |    ~113s    |    **~3s**  |